Fix eventtype Error

+++ src/main/java/io/lolyay/jlavalink/v4/ws/packet/packets/S2C/S2CStatsPacket.java
@@ -62,9 +62,9 @@ public class S2CStatsPacket extends AbstractPacket implements S2CPacket {
         @Expose
         int cores;
         @Expose
-        long systemLoad;
+        float systemLoad;
         @Expose
-        long lavalinkLoad;
+        float lavalinkLoad;
     }
 
     public static class FrameStats {
